Added server-side validation of category

+++ src/main/java/cf/ac/uk/wrackreport/data/jpa/entities/CategoryEntity.java
@@ -18,7 +18,7 @@ public class CategoryEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "category_id")
-    private Long id;
+    private short id;
 
     @Column(name = "category_name")
     private String name;
